update java version to 8-141.
We had replaced jdk_oracle with oracle_java in ccf9c29a, but we haven't
cleanup all the reminiscence of jdk_oracle, at least, not in tests.
At that time, the version used in the serverspec test happened to be 131,
matches the one installed with oracle_java, so no failure in test.

In this commit, I have removed more jdk_oracle in config, and updated
the version of java to 1.8.141-b1.

The md5 and the url was taken from:

  https://github.com/antoineco/aco-oracle_java/pull/35/commits/d2f724

Signed-off-by: Kai Xia <kai.xia@shinesolutions.com>